+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- changed: (Fantom) Update explorer links
+- removed: (Fantom) Removed ftmscan from network adaptor config
+
 ## 4.45.1 (2025-04-29)
 
 - fixed: (TRX) Energy estimations based on contract storage state